+++ src/hooks/useProfile.jsx
@@ -2,10 +2,12 @@ import { useEffect, useState, useContext } from 'react';
 import { supabase } from '../supabase/client';
 import { AuthContext } from '../context/AuthProvider';
 import { handleImageChange, handleImageUpload } from '../shared/utils/fileUtils';
-import { validateBlog, validateEmail, validateGithub, validateNickname, validatePassword } from '../shared/utils/validationUtils';
+// import { validateBlog, validateEmail, validateGithub, validateNickname, validatePassword } from '../shared/utils/validationUtils';
+import { useValidation } from './useValidation';
 
 export const useProfile = () => {
   const { isLogin } = useContext(AuthContext);
+  const { errors, validateField, validateForm, setErrors } = useValidation();
   const [profile, setProfile] = useState({
     image_url: "",
     nickname: "",
@@ -17,7 +19,7 @@ export const useProfile = () => {
   const [image, setImage] = useState(null);
   const [selectedInterests, setSelectedInterests] = useState([]);
   const [preview, setPreview] = useState(null);
-  const [errors, setErrors] = useState({});
+  // const [errors, setErrors] = useState({});
 
   useEffect(() => {
     if (!isLogin) return;
@@ -159,34 +161,34 @@ export const useProfile = () => {
     }));
   };
 
-  // 전체 검증 
-  const validateForm = () => {
-    const newErrors = {};
-    Object.keys(profile).forEach((key) => {
-      const error = validateField(key, profile[key]);
-      if (error) newErrors[key] = error;
-    });
-    return newErrors;
-  };
+  // // 전체 검증 
+  // const validateForm = () => {
+  //   const newErrors = {};
+  //   Object.keys(profile).forEach((key) => {
+  //     const error = validateField(key, profile[key]);
+  //     if (error) newErrors[key] = error;
+  //   });
+  //   return newErrors;
+  // };
 
-  // 입력값 검증
-  const validateField = (name, value) => {
-    switch (name) {
-      case "email":
-        return validateEmail(value) ? "" : "이메일 형식이 올바르지 않습니다.";
-      case "nickname":
-        return validateNickname(value) ? "" : "닉네임은 2~8자 한글, 영어, 숫자 조합만 가능합니다.";
-      case "github":
-        return validateGithub(value) ? "" : "GitHub URL 형식이 올바르지 않습니다.";
-      case "blog":
-        return validateBlog(value) ? "" : "블로그 URL 형식이 올바르지 않습니다.";
-      case "password":
-        if(!value) return "";
-        return validatePassword(value) ? "" : "비밀번호는 소문자, 숫자, 특수문자 포함하여 8자 이상";
-      default:
-        return "";
-    }
-  };
+  // // 입력값 검증
+  // const validateField = (name, value) => {
+  //   switch (name) {
+  //     case "email":
+  //       return validateEmail(value) ? "" : "이메일 형식이 올바르지 않습니다.";
+  //     case "nickname":
+  //       return validateNickname(value) ? "" : "닉네임은 2~8자 한글, 영어, 숫자 조합만 가능합니다.";
+  //     case "github":
+  //       return validateGithub(value) ? "" : "GitHub URL 형식이 올바르지 않습니다.";
+  //     case "blog":
+  //       return validateBlog(value) ? "" : "블로그 URL 형식이 올바르지 않습니다.";
+  //     case "password":
+  //       if(!value) return "";
+  //       return validatePassword(value) ? "" : "비밀번호는 소문자, 숫자, 특수문자 포함하여 8자 이상";
+  //     default:
+  //       return "";
+  //   }
+  // };
   
   return {
     profile,
